Chapter-3
Adding an edit & add screen to the app. App has two screens one for
displaying and iterating through cars and now one to add or edit cars.
The newest screen the edit/add screen is set up using IB but needs to
be hooked up properly to the original screen.